+++ lib/mailchimp_api/collection_parsers/campaign_folder.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module MailchimpAPI::CollectionParsers
+  class CampaignFolder < Base
+    def element_key
+      'folders'
+    end
+  end
+end