+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## [1.4.0]
+
+- Support of `/file-manager/folders`
+
+## [1.3.0]
+
+- Support of `/campaign-folders`
+
 ## [1.2.0]
 
 - Support of `/campaigns`